fix mask

+++ ltr/models/depth_segm/depth_segm_attention_net1.py
@@ -425,7 +425,7 @@ class DepthSegmNetAttention01(nn.Module):
         mask = 1 - F.interpolate(mask_train[0], size=(f_train_rgb.shape[-2], f_train_rgb.shape[-1])) # [B,1,384, 384] -> [B,1,24,24]
         mask = torch.cat((mask, mask), dim=2)                                                        # Bx1x24x24  + Bx1x24x24  -> Bx1x(24+24)x24
         mask = self.mask_embedding(mask).view(mask.shape[0], -1)                                  # Bx1x6x6 -> Bx1xPatches
-        mask = (mask>0.5).float()                                                                    # used to ignore bg patches in key
+        mask = mask>0.5                                                                    # used to ignore bg patches in key
 
         out, attn_weights3 = self.cross_attn(template, search_region, key_padding_mask=mask) # B x Patches x C [rgb + d ]
         n_patches = out.shape[1]                                                             # RGB + D patches
